+++ apps/tickets/models/__init__.py
@@ -2,3 +2,5 @@
 #
 from .ticket import *
 from .comment import *
+from .flow import *
+
