+++ apps/tickets/signals_handler/ticket.py
@@ -3,7 +3,7 @@
 from django.dispatch import receiver
 
 from common.utils import get_logger
-from tickets.models import Ticket, TicketFlowApprove
+from tickets.models import Ticket, TicketFlowApprovalRule
 from ..signals import post_change_ticket_action, post_or_update_change_template_approve
 
 logger = get_logger(__name__)
@@ -14,7 +14,7 @@ def on_post_change_ticket_action(sender, ticket, action, **kwargs):
     ticket.handler.dispatch(action)
 
 
-@receiver(post_or_update_change_template_approve, sender=TicketFlowApprove)
+@receiver(post_or_update_change_template_approve, sender=TicketFlowApprovalRule)
 def post_or_update_change_template_approve(sender, qs, **kwargs):
     updates = []
     for instance in qs:
